requirements.txt
flask
flask-restful
flask-cors
pandas
scikit-learn
scikit-surprise
numpy
joblib
ascii_graph